backend/app/services/aliyun_tts.py
from typing import Optional
import hashlib
import time

from app.core.config import settings


async def text_to_speech(text: str, voice: str = "en-US-female") -> Optional[str]:
    """
    使用阿里云 TTS 合成语音

    注意：这里是占位实现，实际需要：
    1. 调用阿里云智能语音交互 API
    2. 将音频上传到 OSS
    3. 返回音频 URL

    完整实现需要：
    - 安装 aliyun-python-sdk-core
    - 配置阿里云 TTS 服务
    """
    # TODO: 实现阿里云 TTS 调用
    # 临时返回占位 URL

    # 生成唯一标识
    text_hash = hashlib.md5(text.encode()).hexdigest()[:8]
    timestamp = int(time.time())

    # 占位 URL（实际需要替换为真实 TTS 服务）
    placeholder_url = f"https://tts.placeholder.com/audio/{text_hash}_{timestamp}.mp3"

    return placeholder_url


# 实际实现示例代码（供参考）
"""
from alibabacloud_nls20190201.client import Client as NlsClient
from alibabacloud_tea_openapi import models as open_api_models

def create_tts_client():
    config = open_api_models.Config(
        access_key_id=settings.ALIYUN_ACCESS_KEY_ID,
        access_key_secret=settings.ALIYUN_ACCESS_KEY_SECRET
    )
    config.endpoint = 'nls-meta.cn-shanghai.aliyuncs.com'
    return NlsClient(config)

async def text_to_speech_real(text: str, voice: str) -> Optional[str]:
    client = create_tts_client()
    # ... 调用 TTS API
    # ... 上传音频到 OSS
    # ... 返回 URL
"""
