backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Services
openai==1.10.0
dashscope==1.14.1

# Utilities
python-dotenv==1.0.0
Pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
